Move selection out of markdown conversion
