+++ src/utils/date-utils.ts
@@ -1,0 +1,5 @@
+import dayjs from 'dayjs';
+
+export function convertDate(date: Date): string {
+	return dayjs(date).format('YYYY-MM-DD');
+}
