+++ src/utils/filters.ts
@@ -263,7 +263,7 @@ export const filters: { [key: string]: FilterFunction } = {
 	markdown: (str: string, url?: string): string => {
 		const baseUrl = url || 'about:blank';
 		try {
-			return createMarkdownContent(str, baseUrl, '', true); // Pass true to skip Readability
+			return createMarkdownContent(str, baseUrl);
 		} catch (error) {
 			console.error('Error in createMarkdownContent:', error);
 			return str;
